--- src/screens/Login/LoginPage.module.css
@@ -1,92 +1,0 @@
-.root {
-    width: calc(100% + 32px);
-    height: calc(100% + 32px);
-    display: flex;
-    flex-direction: row;
-    margin: -16px;
-}
-
-.welcomeText {
-    font-size: 32px;
-    font-weight: bold;
-    text-align: center;
-    margin-bottom: 0;
-}
-
-.infoText {
-    font-size: 20px;
-    font-weight: bolder;
-}
-
-.form {
-    flex: 2 0;
-    display: flex;
-    flex-direction: column;
-    padding: 32px;
-    background: #ffffff;
-}
-
-.inputContainer {
-    display: flex;
-    flex-direction: column;
-    padding: 10px 10px 8px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    margin-bottom: 20px;
-    transition: 0.3s;
-}
-
-.inputLabel {
-    font-size: 11px;
-    text-transform: uppercase;
-    font-weight: 600;
-    letter-spacing: 0.7px;
-    color: #8c7569;
-    transition: 0.3s;
-}
-
-.input {
-    outline: 0;
-    border: 0;
-    padding: 4px 0 0;
-    font-size: 14px;
-}
-
-.bttn {
-    align-self: flex-end;
-    padding: 8px 16px;
-    outline: none;
-    color: #ffffff;
-    background-color: #4f6474;
-    cursor: pointer;
-    border: 0;
-    border-radius: 4px;
-    transition: 0.3s;
-    font-size: 24px;
-}
-
-.bttn:hover {
-    background: #122f3d;
-}
-
-.image {
-    flex: 4 0;
-    background-image: url("https://source.unsplash.com/collection/4695781");
-    background-size: cover;
-    background-position: right;
-}
-
-.centerText {
-    text-align: center;
-}
-
-.buttonText {
-    margin-left: 8px;
-    text-decoration: underline;
-    cursor: pointer;
-}
-
-.switchText {
-    margin-top: auto;
-    text-align: center;
-}
