+++ src/screens/admin/Login/LoginPage.jsx
@@ -1,0 +1,63 @@
+import React from "react";
+// import Router from "next/router";
+// import { login } from "../../../actions/User";
+// import urls from "../../../../utils/urls";
+import classes from "./LoginPage.module.css";
+
+const LoginPage = () => {
+  const [email, setEmail] = React.useState("");
+  const [password, setPassword] = React.useState("");
+
+  const handleSubmit = (event) => {
+    event.preventDefault();
+
+    console.log(email);
+    console.log(password);
+
+    // return login(email, password)
+    //   .then(() => Router.replace(urls.pages.adminHome))
+    //   .catch((error) => window.alert(error.message));
+  };
+
+  return (
+    <div className={classes.background}>
+      <div className={classes.formContainer}>
+        <h2 className={classes.welcomeText}>Welcome back</h2>
+        <h6>This is for admin login only</h6>
+        <form className={classes.form} onSubmit={handleSubmit}>
+          <div className={classes.inputContainer}>
+            <label htmlFor="email" className={classes.inputLabel}>
+              Email address
+            </label>
+            <input
+              className={classes.input}
+              required
+              id="email"
+              type="email"
+              value={email}
+              onChange={(event) => setEmail(event.target.value)}
+            />
+          </div>
+          <div className={classes.inputContainer}>
+            <label htmlFor="password" className={classes.inputLabel}>
+              Password
+            </label>
+            <input
+              className={classes.input}
+              required
+              id="password"
+              type="password"
+              value={password}
+              onChange={(event) => setPassword(event.target.value)}
+            />
+          </div>
+          <button className={classes.button} type="submit">
+            Log In
+          </button>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default LoginPage;
