+++ src/screens/admin/Login/LoginPage.module.css
@@ -1,0 +1,76 @@
+.background {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: calc(100% + 32px);
+    height: calc(100% + 32px);
+    margin: -16px;
+    background-image: url("https://source.unsplash.com/collection/4695781");
+    background-size: cover;
+}
+
+.formContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background: white;
+    width: 450px;
+    height: auto;
+    padding: 28px;
+}
+
+.welcomeText {
+    font-size: 24px;
+    font-weight: bold;
+    margin: 0px;
+}
+
+.form {
+    box-sizing: border-box;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+    height: auto;
+    background-color: whitesmoke;
+    padding: 60px;
+}
+
+.inputContainer {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+    margin-bottom: 20px;
+}
+
+.inputLabel {
+    flex: 3;
+    white-space: nowrap;
+    font-size: 14px;
+}
+
+.input {
+    flex: 5;
+    outline: 0;
+    border: whitesmoke solid;
+    padding: 10px;
+    font-size: 12px;
+}
+
+.button {
+    width: 100%;
+    padding: 8px 0px;
+    outline: none;
+    color: #ffffff;
+    background-color: silver;
+    cursor: pointer;
+    border: 0;
+    border-radius: 4px;
+    transition: 0.3s;
+    font-size: 14px;
+}
+
+.button:hover {
+    background: grey;
+}
